clean up inputs
