+++ apps/server/src/modules/user/controller/user.controller.ts
@@ -1,8 +1,9 @@
 import { ApiTags } from '@nestjs/swagger';
 
 import { Controller, Get } from '@nestjs/common';
+import { ICurrentUser } from '@shared/domain';
+
 import { Authenticate, CurrentUser } from '../../authentication/decorator/auth.decorator';
-import { ICurrentUser } from '../../authentication/interface/jwt-payload';
 import { UserUC } from '../uc';
 import { ResolvedUser } from './dto/ResolvedUser.dto';
