+++ apps/server/src/shared/domain/interface/index.ts
@@ -1,1 +1,2 @@
 export * from './find-options';
+export * from './user';
